+++ crates/tidy-browser/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Subcommand: binary-cookies
 - Output format json, jsonlines
+- Generate shell completions
 
 ## [0.2.0] - 2025-07-27